refactor: Divisão de responsabilidade da classe ProdutoValidacao
